// src/types/photoLog.ts
export interface PhotoLogRow {
  project_uuid?: string;
  group_date?: string;
  photo_log_uuid?: string;
  photo_order?: string;
  created_at?: string;
  photo_date?: string;
  photo_description?: string;
  photo_uuid?: string;
  photo_name?: string;
  photo_size?: string;
  photo_type?: string;
  source_view_url?: string;
  source_download_url?: string;
  preview_thumbnail_url?: string;
  preview_large_url?: string;
}